Extract GPUI API primitives into a new gpui_types crate
Introduce crates/gpui_types, holding the self-contained value types shared by GPUI and its platform backends. In this port it contains geometry (Pixels, Point, Size, Bounds, ...) and the taffy layout bridge (AvailableSpace, ...).

gpui now depends on and re-exports gpui_types, so gpui::Pixels and gpui_types::Pixels name the same type and downstream code is unchanged. gpui_linux imports its geometry primitives from gpui_types directly.

Unlike upstream, color is not extracted: gpui-ce's color module re-exports palette's Hsla/Rgba, and its Background/LinearColorStop are backed by SceneHsla (defined in scene.rs), so that module is not self-contained.

Methods that require App/Window stay in gpui as the BoundsExt and HasRemSize extension traits. Conversions bridging two foreign types (taffy, lyon) moved to gpui_types to satisfy the orphan rule, with the lyon conversions behind an optional lyon feature.

Release Notes:

- N/A
