+++ crates/gpui/src/prelude.rs
@@ -3,7 +3,7 @@
 //! application to avoid having to import each trait individually.
 
 pub use crate::{
-    AppContext as _, BorrowAppContext, Context, Element, InteractiveElement, IntoElement,
-    ParentElement, Refineable, Render, RenderOnce, StatefulInteractiveElement, Styled, StyledImage,
-    TaskExt as _, VisualContext, util::FluentBuilder,
+    AppContext as _, BorrowAppContext, BoundsExt, Context, Element, InteractiveElement,
+    IntoElement, ParentElement, Refineable, Render, RenderOnce, StatefulInteractiveElement, Styled,
+    StyledImage, TaskExt as _, VisualContext, util::FluentBuilder,
 };
